Adjust verifycmd for MUSCLE
Without any command-line arguments, the behavior and exit code of the MUSCLE executable
depend on whether isatty(stdin). If that's not the case (as in certain unattended
environments, but evidently not Travis), it'll try to read from stdin and fail. Very
nasty to debug because it then works as expected when you try it interactively! Running
with -version is more predictable.