fix to mafft verify command, call command
The mafft binary is called by a shell script that comes with the mafft
package. This shell script relies on $pwd being the mafft package
directory.  Differences in environment can change the subprocess pwd,
thus it is necessary to cd to the mafft directory for consistency so
calls to mafft from python work as expected across different platforms