fixes in response to feedback
—local pair and —global pair are now mutually exclusive via argparse
add_mutually_exclusive_group().

The —maxiter argument note now mentions that the default changes to
1000 if —localpair or —globalpair is specified.

The  —threads argument now defaults to “-1” to use all available cores
(this is the value MAFFT looks for to use all core), and the note
reflects this. A check has been added for zero or other negative values.

in mafft.py files are now read and written as text, rather than binary
as they were previously.