NCBI fetch functions now permit chunking for batch downloading
NCBI fetch functions now permit chunking for batch downloading. The
default chunk size is 1 so that the behavior is the same as before. If
a chunk size is specified, all accessions in a chunk are downloaded
together and concatenated into a common file. If —combinedFilePrefix is
specified, the chunks are then themselves concatenated together (as
before). The max chunk size is 500 as per the NCBI examples, and
Biopython handles rate limiting to 3/sec as per the NCBI guidelines. If
a list of accessions is specified and chunkSize is >1, the output
filename for the chunk is {accessionList}.{fileExt} unless the filename
would be >254 characters long, then it is “chunk-{chunkNum}.{fileExt}”