changed lastal db build rule so that the lastal db is built from a file of accessions, if specified in the config.
This paves the way for using the pan-viral database available from the
NCBI here:
http://www.ncbi.nlm.nih.gov/genomes/GenomesGroup.cgi?taxid=10239

Either via a snapshot of the accession list files available for
download, or via a a download of the latest file. Functions have been
added to common.rules, and util.file to download a file from a URI, or
to read lines from a text file specified by URI

Added a variable to config.json to hold the file path (or URL?) of the
accession list used for building the lastal db, and changed the lastal
db path to be different from that of the ref genome.