Make ansible vagrant easier to test changes
Previously easy_deployed relied on git clone of repo from github, which
makes playbook changes hard to test without committing to the master
repo.

The clone is still the default behavior, but now we have two other
options that can be toggled via the --extra-vars=deploy=archive/sync
command line options passed to ansible playbook.

For archive, we vagrant sync the local git repo into the VM, then perform a git
archive export to the final project directory so tool installation in VM
does not affect the host machine.

For sync, the VM's installation of viral-ngs is directly the hosts
viral-ngs repo so all changes are synced accordingly. This is the
fastest for iteration but least portable option.