fix to snpeff availabl_databases() for conda-based installs
the version of snpeff installed by conda relies on a shell script
wrapper, and must be called differently than a standard jar file. this
difference in invocation was present in snpeff.execute() but not
snpeff.available_databases(). The latter now relies on
snpeff.execute(). The execute() function itself has been changed to use
util.misc.run() rather than subprocess.check_call() so we can capture
the output (it is read to get the list of available databases)