fix bug introduced by previous
Fixes an issue caused by the last commit. annotate_vcf() was passing an
open file handle as stdout to snpeff.execute(), which did not work
following the change from subprocess.check_call() to util.misc.run() in
execute(). The run() output is now decoded and written to the open file
handle