guess version number when git is unavailable or source is not git checkout
add a function to util/version.py, approx_version_number(), to output a
version number in the cases where git is unavailable, the copy of
viral-ngs is not a git checkout, and a VERSION file is not present. It
first tries to parse a version number from the viral-ngs root path
(likely for copies extracted from tagged releases), then falls back to
the mtime of the version.py file (since this will reflect either the
commit or tag date when the archive was created). In cases where mtime
cannot be accessed, the current time is used as the version number