Upgrade util.misc.run more like py3.5 run.
Change the run function to be more feature complete like py3.5's run
function. Relies on temporary files for stdout/stderr since
subprocess.PIPE is not allowed when we're trying to capture them.
Otherwise, use check_output like before which is more efficient if we
are redirecting stderr to stdout or not capturing.

Therefore allow kraken to separate stdout as datastream writing to file
and stderr in log messages.