Fix metagenomics input and provide more resources
The metagenomics tools took as input the file "{sample}.bam" in the
source subdir, but this file is not created when demultiplexing raw
data and the snakemake pipeline is not setup to create this file.
Thus, snakemake is unable to find input for the metagenomics tools.
This fix changes the input to be "{sample}.raw.bam" in the per_sample
subdir. The snakemake pipeline is able to create this file even
when demultiplexing raw reads, so it is able to create input for
the metagenomics tasks.

This also gives the metagenomics tools more resources; it requests
more memory and a queue with longer runtime.