Add read depth to merged intrahost variation output
This adds the 'DP' field to the VCF file created at the end of
the intrahost component of the pipeline. The field gives, at
each position, the read depth for each sample. Together with
the allele frequency ('AF'), this can provide allele counts at
each position in each sample. This can be useful, e.g., for
statistical tests that compare the abundance of an iSNV between
samples.