Fix JSON serializability bug
When passing information into params, snakemake requires that it be
JSON serializable. However, `read_samples_file` is a generator and
is not serializable. snakemake fails with `TypeError: <generator
object read_samples_file at ... is not JSON serializable>`. This
fix wraps the call to `read_samples_file` with list(..) to obtain
the output of the generator, which can be serialized.