Correct passing of samples list to merge_to_vcf command
The list of samples was passed as an argument to the `merge_to_vcf`
command of `intrahost.py` as single string representing the list.
The result was that `merge_to_vcf` could not properly recognize
which samples were input. This commit changes the snakemake rule to
pass each sample name as a separate argument to the command, as they
are intended to be.