Fix various issues in interhost rules
This fixes issues in the interhost rules that prevented targets
from running. Now, the `all_ref_guided` target runs successfully,
along with all the others it invokes.

The fix in `assembly.py` is needed for segmented genomes. Without
this fix, the fasta file would give the same header for each segment
and downstream steps would fail because of this.