SampleSheet reader now uses csv.reader() rather than str.split()
Previously, commas in sample names could cause problems when reading in
sample sheets, since the reader was naively splitting on commas. This
incorporates changes to uses Python’s csv.reader() to parse
SampleSheet.csv rows, with a custom dialect to support escaping quotes
with backslash characters. It also adds a unit test with a SampleSheet
that was problematic prior to these changes