Use unpaired trimmomatic output in downstream assembly steps
Trimmomatic can output single-end reads if mates are dropped.
Previously, its single-end output was stored in tmp files and trashed.
This change keeps its unpaired output and allows those reads to be
used as input for Trinity assembly.

This is particularly useful for low quality and quantity samples. In
these samples, often the second mate of a pair has a drop in quality
(at least on MiSeq sequencing) and is tossed by Trimmomatic; this
change would allow the first mate of the pair to be used in assembly.