perform pre-Trinity subsampling in bam space
Perform pre-Trinity subsampling in bam space. If the number of paired
reads (after purge, or not) is lower than the Trinity subsampling
threshold, subsample the unpaired reads to try to reach the total.
Otherwise, if the post-purge paired read count is greater than the
Trinity input, downsample to reach it. A few small improvements to the
samtools and Picard downsample functions.