Run fastqc on raw, cleaned, taxfilt, and align-to-self reads
This modifies snakemake rules so that the `all_reports` rule
runs fastqc on raw reads, cleaned reads, taxonomic filtered
reads, and reads aligned to a sample's own assembly. (Prior
to this change, only the latter reads were input to fastqc.)

Additionally, this adds an `all_ref_guided_fastqc` rule that
runs fastqc on reads aligned to the reference.