easy-deploy-script.sh: shorten PS1; reduce size memory allocation test
This shortens the shell prompt by modifying PS1 after the environment
has been activated. It also reduces the memory allocation check amount
to 768MB to account for the UGER shift from m_mem_free to h_vmem. The
script can still tell if it is on an interactive node or a login node,
while providing enough RAM to function.