avoid RevertSam in deplete_human
RevertSam is expensive, since it reads, alters, and sorts all records
in a bam file. This changes main_deplete_human so RevertSam is only
executed if the input bam file is aligned. It determines if the file is
aligned by checking whether the @SQ line exists and is populated in the
bam header (via pysam). Since RevertSam also sanitizes, we lose some
functionality: “Reads discarded include (but are not limited to) paired
reads with missing mates, duplicated records, records with mismatches
in length of bases and qualities.”

revertBam has been made an optional argument via nargs=‘?’, so calls to
deplete_human will not need to be changed. A test case has been added
for aligned input