add SamtoolsTool.filterByCigarString(); filter v-phaser input to remove leading and trailing indels
Add a function, filterByCigarString(), that uses two samtools processes
to filter a bam file based on a regex applied to each aligned read’s
cigar string. The initial use case (included in this commit) is to
remove reads with leading and trailing indels, which would otherwise
cause v-phaser2 to crash. A unit test is also added for
SamtoolsTool.filterByCigarString().