harden v-phaser2 against empty input
It is possible for V-Phaser2 input to contain zero reads after removing
doubly-mapped reads, in the case of low-quality data. This hardens the
vphaser step to be tolerant of such cases by writing out a blank
v-phaser2 output file. An integration test is also added, and the
test-skip decorator for vphaser under osx is commented out.