work around tbl2asn out-of-date quirk
tbl2asn has a fun quirk where if the build is more than a year old it
exits with a non-zero code and tells you to upgrade. we attempt to work
around that by examining its output. also add tbl2asn to
requirements-conda.txt and omit the version pin for tbl2asn so we
always install the latest version available on bioconda