Set a threshold on alignment score when generating spike-in reports (#666)
The spike-in reports uses bwa to align reads to spike-in sequences,
but bwa does not set a threshold on alignment score when reporting
alignments given paired-end data. As a result, many spike-ins can
be reported as hits but are false positives (for example, some
spike-in sequences have a polyA tail and reads with a polyA stretch
can map to these spike-in sequences). Setting a threshold on which
alignments are reported alleviates this problem.