replace deprecated universal newline 'U' with py3-compatible
Using the mode ‘U’ to tell Python to open text files with universal
newline support is deprecated in py3, and may be removed in the future.
This commit alters open_or_gzopen to set newline=None (universal
newline) under py3 when ‘U’ is given in the mode, and changes a few
open() calls to use open_or_gzopen(). Add a few test cases for
tab-delimited sample sheets with various line endings. Also replace the
deprecated “py.test” with “pytest”.