docker improvements (#711)
This updates the viral-baseimage to one that supports gsutil backed with crcmod (earlier versions neglected to install this apt package). This also fixes the source activate invocation in env_wrapper.sh (the Docker image's entrypoint) so that it no longer clobbers certain binaries in the conda environment with recursive symlinks.