Create a stub CondaPackage for use with tests.
For all tests except test_tools, stub out the CondaPackage class using a globally autoused conftest
fixture. This CondaPackage doesn't do anything and simply assumes that the packages are installed.
This is useful because simple commands like "conda list" take considerable amounts of time. The
conda commands are still tested in test_tools, which overrides the stub_conda fixture with a local one.