let scaffolder use ambiguous alignments when no unambiguous ones exist (#904)
When scaffolding contigs to a reference, if no unambiguous alignments cover a part of reference, let the scaffolder consider ambiguous ones, when all alignments mostly agree on the sequence.  The ambiguity can be the result of a misassembly, where the same reference part appears twice in a contig.   Also, in case of a repeat, dropping all ambiguous alignments could leave a hole in the assembly.