roll back test input to state of origin/master; add functions to util.file to sanitize fasta file IDs
fasta ID is sanitized for picard CreateSequenceDictionary to adhere to character set restrictions in SAM/BAM RNAME spec. see: https://github.com/samtools/hts-specs/pull/333

Roll back test input now that we are doing the sanitization. As of this commit, sanitization is only performed on fasta files destined for CreateSequenceDictionary