maintain RG info in clumpify dedup; move processing to bbmap.py
for bbmap clumpify de-dup, merge like-library RGs and perform deduplication on each, then gather the IDs of kept reads, and filter the input sam based on the list of IDs to keep so as to maintain header and RG information. move most of the theprocessing to bbmap.py::dedup_clumpify so it has a more simple interface that accepts one bam and emits one bam. ToDo: parallelize across LBs