do not discard read IDs for single-end reads after running mvicuna
expects read IDs to have a /1 mate suffix, and only includes those that do, which single-end reads do not have (nor interleaved fastqs, but that's a separate issue and we're generating the fastqs for this internal function ourselves so interleaving should not be an issue). So no single-end IDs were preserved. Now they are. A separate issue is that mvicuna does not de-duplicate single-end reads, so the mvicuna step is purely a pass-through.