Update read_utils.py
The old mvicuna post-processing code expects read IDs to have a /1 mate suffix, and only includes those that do, which single-end reads do not have (nor interleaved fastqs, but that's a separate issue and we're generating the fastqs for this internal function ourselves so interleaving should not be an issue). So no single-end IDs were preserved. This ensures single-end read IDs make it through this post-processing.