in read_utils.align_and_fix(), filter input and output via samtools.filter_to_mapped_reads()
in read_utils.align_and_fix(), filter input and output via samtools.filter_to_mapped_reads(), such that marked pcr duplicates and mapped singletons are removed on the input side, and on the output side marked duplicates are removed as well as any reads not properly paired, except for single-end reads, which are allowed through if mapped (or in any case if allow_unmapped=True). This is intended to address a specific issue where NovaSeq contamination presented as singleton reads.