rewrite main_flowcell_metadata() to return "UNKNOWN" rather than error if sequencer model could not be determined
rewrite main_flowcell_metadata() to return "UNKNOWN" rather than error if sequencer model could not be determined based on flowcell ID or tile count. Prefer flowcell ID, falling back to tile count. If no matches for either, say the sequencer model is unknown. This also adds unit tests for instances where the fcid ID is novel but the tile count is known (ex. new series of flowcell), for when the flowcell id is known but the tile count is novel (this should not happen, but we defer model guess to the flowcell ID), and where both flowcell ID and tile count are unknown (ex. completely new machine). This also exposes `RunInfo.get_section_count()` since some machines (like the NextSeq 550) use it to calculate total tile count.