WIP add collapsing of barcode or barcode pair duplicates as part of illumina.py::illumina_demux() via --collapse_duplicated_barcodes
add collapsing of barcode or barcode pair duplicates as part of `illumina.py::main_illumina_demux()` via `--collapse_duplicated_barcodes` flag. The deduplication is done on the basis of (`barcode_1`) for single-end Illumina indexing, or (`barcode_1`,`barcode_2`) for paired-end Illumina indexing, with each group of duplicate indexes collapsing to a single row prior to demultiplexing via picard IlluminaBasecallsToSam. The deduplication happens within the `SampleSheet()` class after reading in the samplesheet (with future format flexibility in mind). Within each deduplicated group, the `barcode_[1-2]` columns are passed through verbatim.  Other columns in the sample sheet are either reported verbatim if only one distinct value exists within each deduplicated group, or set to a single string that represents all values within each group. The single string can be the delimited join of member values, optionally followed by a supplied suffix (`_muxed` is the default), or if the length of such a constructed string exceeds a specified limit, the single string is set to a truncated md5 hash of joined member values (without delimiting characters) followed by an optional suffix. This deduplication occurs in `util/misc.py::collapse_dup_strs_to_str_or_md5()`. The intent is for the new column values for each deduplicated group to reflect the values from the column prior to deduplication, so changes in sample sheet construction upstream manifest in a way that allows for (some) comparison between deduplicated sample sheets. In the case of a mixed pool where some barcodes or barcode pairs are deduplicated but other barcode pairs are distinct, the distinct rows are passed through without modification.